Add reset button to more settings
