+++ webpages/settings/index.html
@@ -377,6 +377,9 @@
               <div class="export-actions-group">
                 <button class="large-button" @click="viewSettings()">{{ msg("viewSettings") }}</button>
               </div>
+              <div class="export-actions-group">
+                <button class="large-button danger" @click="resetSettings()">{{ msg("resetSettings") }}</button>
+              </div>
             </div>
           </div>
         </div>
